% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count.genus}
\alias{count.genus}
\title{Genus-level taxa count data for 574 observations.}
\format{
A 574 x 133 Matrix
\describe{
\item{rownames}{sample id}
\item{colnames}{generic name for each taxa conut }
}
}
\usage{
count.genus
}
\description{
A dataset containing genus-level count data pooled from five studies
}
\keyword{datasets}
