% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.meta}
\alias{data.meta}
\title{The data are derived from several real gut microbiome studies.}
\format{
A list with 3 elements
\describe{
\item{OTU}{a list contains 5 OTU counts}
\item{Tax}{a matrix of taxonomy table from Rank1 (kingdom level) to Rank6 (genus level)}
\item{covariate}{covariates of interest}
}
}
\usage{
data.meta
}
\description{
The data are derived from several real gut microbiome studies.
}
\keyword{datasets}
