% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeta.R
\name{.resample.work.one.meta}
\alias{.resample.work.one.meta}
\title{Title}
\usage{
.resample.work.one.meta(
  X.list,
  X.par.index,
  n.par.interest.beta,
  col.index.list,
  score.stat.meta,
  S.beta.list.meta,
  I.beta.list.meta,
  start.nperm,
  end.nperm,
  n.one,
  one.acc,
  Method = "MV",
  W = NULL
)
}
\arguments{
\item{W}{}
}
\description{
Title
}
