% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Subject-level data for 574 observations.}
\format{
A dataframe
\describe{
\item{Sample_id}{Sample id for each observation}
\item{External_id}{External_id}
\item{Age}{Age for each observation}
\item{Gender}{Gender, F or M}
\item{BMI}{BMI for each observation}
...
\item{Group}{Denote each group belongs to case or control group}
...
}
}
\usage{
meta
}
\description{
A dataset containing subject-level variables pooled from five studies of colon cancer status
}
\keyword{datasets}
