% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeta_GEE.R
\name{QCAT_GEE_Meta}
\alias{QCAT_GEE_Meta}
\title{Title}
\usage{
QCAT_GEE_Meta(
  OTU,
  Z,
  Z.index,
  Method = "FE-MV",
  min.depth = 0,
  n.perm = NULL,
  Tax = NULL,
  fdr.alpha = 0.05
)
}
\arguments{
\item{fdr.alpha}{}
}
\description{
Title
}
