% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeta.R
\name{QCAT_Meta}
\alias{QCAT_Meta}
\title{Title}
\usage{
QCAT_Meta(
  OTU,
  X,
  X.index,
  Tax = NULL,
  Method = "MV",
  Weight = NULL,
  min.depth = 0,
  n.perm = NULL,
  fdr.alpha = 0.05
)
}
\arguments{
\item{fdr.alpha}{}
}
\description{
Title
}
