% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax}
\alias{tax}
\title{Taxonomy table}
\format{
A dataframe
\describe{
\item{OTUID}{id}
\item{kingdom}{the scientific name for the kingdom that the taxa belongs}
\item{phylum}{the scientific name for the phylum that the taxa belongs}
\item{class}{the scientific name for the class that the taxa belongs}
\item{order}{the scientific name for the order that the taxa belongs}
\item{family}{the scientific name for the family that the taxa belongs}
\item{genus}{generic name for each taxa conut}
\item{mOTU}{mOTU}
}
}
\usage{
tax
}
\description{
Taxonomy ranks for the taxa counts in count.genus data
}
\keyword{datasets}
