% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeta.R
\name{QCAT_Meta}
\alias{QCAT_Meta}
\title{Title}
\usage{
QCAT_Meta(
  OTU,
  X,
  X.index,
  Tax = NULL,
  Method = "FE-MV",
  min.depth = 0,
  n.perm = NULL,
  fdr.alpha = 0.05
)
}
\arguments{
\item{OTU}{a list of matrices contains counts with each row corresponds to a sample and each column corresponds to an OTU or a taxa. Each matrices' taxas are better to the same. Column name is mandatory.}

\item{X}{a list of matrices contains covariates for the positive-part test with each column pertains to one variable (pertains to the covariate of interest or the confounders). The number of elements of of X and OTU must be the same. The column number of each matrix in this list must be the same.}

\item{X.index}{a vector indicate the columns in X for the covariate(s) of interest.}

\item{Tax}{a matrix define the taxonomy ranks with each row corresponds to an OTU or a taxa and each column corresponds to a rank (start from the higher taxonomic level). Row name is mandatory and should be consistent with the column name of the OTU table,  Column name should be formated as "Rank1", "Rank2 ()"... etc
If provided, tests will be performed for lineages based on the taxonomic rank. The output contains P-values for all lineages; a list of significant lineages controlling the false discovery rate (based on resampling p-value if resampling test was performed); p-values of the global tests (Fisher- and Simes-combined the p-values for testing lineages).
If not provided, one test will be performed with all the OTUs and one p-value will be output.}

\item{Method}{The meta-analysis method to be used. Including fixed effect methods such as FE-MV test and FE-VC test and random effect methods like Het-SKAT and RE-SKAT test.}

\item{min.depth}{keep samples with depths >= min.depth.}

\item{n.perm}{perform asymptotic test is n.resample is null, other perform resampling tests using the specified number of resamplings.}

\item{fdr.alpha}{false discovery rate for multiple tests on the lineages.}
}
\value{
A list with this elements
\item{lineage.pval}{p-values for all lineages. By default ( Method = "FE-MV", n.perm = NULL ), only the asymptotic test will be performed. If use random effect meta-analysis methods ( Method = "RE-SKAT" or Method = "Het-SKAT" ), then resampling test must be performed.}
\item{sig.lineage}{a vector of significant lineages.}
}
\description{
Title
}
\examples{
data(data.meta)
OTU = data.meta$OTU
Tax = data.meta$Tax
case = data.meta$covariate
QCAT_Meta(OTU, case, 1, Tax, Method = "FE-MV", min.depth=0, n.perm=NULL, fdr.alpha=0.05)
}
