% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMeta.R
\name{QCAT_Meta}
\alias{QCAT_Meta}
\title{Title}
\usage{
QCAT_Meta(
  OTU,
  X,
  X.index,
  Tax = NULL,
  Method = "FE-MV",
  min.depth = 0,
  n.perm = NULL,
  use.cpp = F,
  fdr.alpha = 0.05
)
}
\arguments{
\item{OTU}{a list of matrices containing OTU counts with each row corresponding to a sample and each column corresponding to an OTU or taxa. Each matrix's taxas are better to the same. The column name is mandatory.}

\item{X}{a list of matrices containing covariates for the positive-part test with each column pertaining to one variable (pertains to the covariate of interest or the confounders). The number of elements of X and OTU must be the same. The column number of each matrix in this list must be the same.}

\item{X.index}{a vector indicate the columns in X for the covariate(s) of interest.}

\item{Tax}{a matrix defines the taxonomy ranks with each row corresponding to an OTU or a taxa and each column corresponding to a rank (starting from the higher taxonomic level). Row name is mandatory and should be consistent with the column name of the OTU table, Column name should be formatted as "Rank1", "Rank2 ()"... etc
If provided, tests will be performed for lineages based on the taxonomic rank. The output contains P-values for all lineages; a list of significant lineages controlling the false discovery rate (based on resampling p-value if resampling test was performed).
If not provided, one test will be performed with all the OTUs and one p-value will be output.}

\item{Method}{Meta-analysis method to be used. Including fixed effect methods such as the FE-MV test and FE-VC test and random effect methods like RE-MV and RE-VC test.}

\item{min.depth}{keep samples with depths >= min.depth.}

\item{n.perm}{perform asymptotic test if n.perm is null, otherwise perform permutation tests using the specified number of resamplings.}

\item{use.cpp}{Logical value (default F). Whether to use Rcpp or not for resampling test.}

\item{fdr.alpha}{false discovery rate for multiple tests on the lineages.}
}
\value{
A list with this elements
\item{lineage.pval}{p-values for all lineages. By default ( Method = "FE-MV", n.perm = NULL ), only the asymptotic test will be performed. If using random effect meta-analysis methods ( Method = "RE-VC" or Method = "RE-MV" ), then resampling test must be performed.}
\item{sig.lineage}{a vector of significant lineages.}
}
\description{
Title
}
\examples{
data(data.meta)
OTU = data.meta$OTU
Tax = data.meta$Tax
case = data.meta$covariate
QCAT_Meta(OTU, case, 1, Tax, Method = "FE-MV", min.depth=0, n.perm=NULL, fdr.alpha=0.05)
}
\references{
Tang ZZ, Chen G, Alekseyenko AV, Li H. (2017) A general framework for association analysis of microbial communities on a taxonomic tree.
\emph{Bioinformatics}
\doi{10.1093/bioinformatics/btw804}.

Lee S, Teslovich TM, Boehnke M, Lin X. (2013) General framework for meta-analysis of rare variants in sequencing association studies. Am J Hum Genet.
\emph{Am J Hum Genet}
\doi{10.1016/j.ajhg.2013.05.010}.

Benjamini, Yoav, and Yosef Hochberg.(1995) Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing.
\emph{Journal of the Royal Statistical Society. Series B}
}
