% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax}
\alias{tax}
\title{Taxonomy table}
\format{
A dataframe
\describe{
\item{OTUID}{id}
\item{kingdom}{the scientific name of the kingdom that the taxa belongs to}
\item{phylum}{the scientific name of the phylum that the taxa belongs to}
\item{class}{the scientific name of the class that the taxa belongs to}
\item{order}{the scientific name of the order that the taxa belongs to}
\item{family}{the scientific name of the family that the taxa belongs to}
\item{genus}{generic name for each taxa conut}
\item{mOTU}{mOTU}
}
}
\usage{
tax
}
\description{
Taxonomy ranks for the taxa counts in count.genus data
}
\keyword{datasets}
